# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Agent": {
        "properties": {
          "harness": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "mcp_servers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "model": {
            "type": "string"
          }
        },
        "required": [
          "harness",
          "model",
          "instructions",
          "mcp_servers"
        ],
        "type": "object"
      },
      "ClaimRequest": {
        "properties": {
          "environment_types": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "harnesses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ClaimResponse": {
        "properties": {
          "session": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Session"
              }
            ],
            "description": "The claimed session, or null if there is no pending work.",
            "nullable": true
          }
        },
        "type": "object"
      },
      "CreateSessionRequest": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment"
          },
          "input": {
            "type": "string"
          }
        },
        "required": [
          "agent",
          "environment",
          "input"
        ],
        "type": "object"
      },
      "EnrollHostRequest": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "EnrollHostResponse": {
        "properties": {
          "host": {
            "$ref": "#/components/schemas/Host"
          },
          "token": {
            "description": "The bearer token for this host. Shown exactly once; store it safely.",
            "type": "string"
          }
        },
        "required": [
          "host",
          "token"
        ],
        "type": "object"
      },
      "Environment": {
        "properties": {
          "base_branch": {
            "description": "Branch to base work on; None = repo default branch",
            "nullable": true,
            "type": "string"
          },
          "kind": {
            "description": "\"self_hosted\" or \"github_repo\"",
            "type": "string"
          },
          "repo": {
            "description": "Required when kind == \"github_repo\": \"owner/name\"",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "FinishRequest": {
        "properties": {
          "error": {
            "nullable": true,
            "type": "string"
          },
          "outcome": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionOutcome"
              }
            ],
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "stop_reason": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "cancel_requested": {
            "type": "boolean"
          }
        },
        "required": [
          "cancel_requested"
        ],
        "type": "object"
      },
      "Host": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "type": "object"
      },
      "IssueCredentialRequest": {
        "properties": {
          "provider": {
            "description": "Provider name, e.g. \"github\"",
            "type": "string"
          }
        },
        "required": [
          "provider"
        ],
        "type": "object"
      },
      "IssueCredentialResponse": {
        "properties": {
          "expires_at": {
            "description": "None for non-expiring credentials",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "secret"
        ],
        "type": "object"
      },
      "ListSessionsResponse": {
        "properties": {
          "next_cursor": {
            "nullable": true,
            "type": "string"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/Session"
            },
            "type": "array"
          }
        },
        "required": [
          "sessions"
        ],
        "type": "object"
      },
      "NewSessionEvent": {
        "properties": {
          "payload": {},
          "seq": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "seq",
          "payload"
        ],
        "type": "object"
      },
      "ReportEventsRequest": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/NewSessionEvent"
            },
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "type": "object"
      },
      "Session": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "cancel_requested": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "finished_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "host_id": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "input": {
            "type": "string"
          },
          "lease_expires_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "outcome": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SessionOutcome"
              }
            ],
            "nullable": true
          },
          "started_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "stop_reason": {
            "nullable": true,
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent",
          "environment",
          "input",
          "status",
          "cancel_requested",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "SessionEvent": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {},
          "seq": {
            "format": "int64",
            "type": "integer"
          },
          "session_id": {
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "seq",
          "payload",
          "created_at"
        ],
        "type": "object"
      },
      "SessionOutcome": {
        "properties": {
          "branch": {
            "nullable": true,
            "type": "string"
          },
          "kind": {
            "description": "\"pr_opened\" | \"pushed_no_pr\" | \"uncommitted_changes\" | \"no_changes\"",
            "type": "string"
          },
          "pr_url": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "SessionStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "type": "string"
      }
    }
  },
  "info": {
    "description": "API for managing Vise agent sessions",
    "license": {
      "name": ""
    },
    "title": "Vise API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/hosts": {
      "get": {
        "operationId": "list_hosts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Host"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List all enrolled hosts"
          }
        },
        "tags": [
          "hosts"
        ]
      },
      "post": {
        "operationId": "enroll_host",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollHostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollHostResponse"
                }
              }
            },
            "description": "Host enrolled; the token is returned exactly once"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/hosts/claim": {
      "post": {
        "operationId": "claim_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            },
            "description": "The claimed session, or null if there is no pending work"
          },
          "401": {
            "description": "Missing or invalid host token"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/hosts/sessions/{id}/credentials": {
      "post": {
        "operationId": "issue_credential",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueCredentialResponse"
                }
              }
            },
            "description": "Short-lived credential for the session"
          },
          "401": {
            "description": "Missing or invalid host token"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "Host no longer holds this session"
          },
          "422": {
            "description": "Provider not applicable to this session"
          },
          "502": {
            "description": "Upstream credential issuer failed"
          },
          "503": {
            "description": "Provider not configured on this server"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/hosts/sessions/{id}/events": {
      "post": {
        "operationId": "report_session_events",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportEventsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Events recorded"
          },
          "401": {
            "description": "Missing or invalid host token"
          },
          "409": {
            "description": "Host no longer holds this session"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/hosts/sessions/{id}/finish": {
      "post": {
        "operationId": "finish_session",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Session finished"
          },
          "401": {
            "description": "Missing or invalid host token"
          },
          "409": {
            "description": "Host no longer holds this session"
          },
          "422": {
            "description": "Status is not terminal"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/hosts/sessions/{id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_session",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Lease extended"
          },
          "401": {
            "description": "Missing or invalid host token"
          },
          "409": {
            "description": "Host no longer holds this session"
          }
        },
        "tags": [
          "hosts"
        ]
      }
    },
    "/sessions": {
      "get": {
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            },
            "description": "List all sessions"
          }
        },
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Session created"
          },
          "422": {
            "description": "Invalid environment"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/{id}": {
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Get a session"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/{id}/cancel": {
      "post": {
        "operationId": "cancel_session",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Cancellation requested (pending sessions are cancelled immediately)"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "Session already finished"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    },
    "/sessions/{id}/events": {
      "get": {
        "operationId": "get_events",
        "parameters": [
          {
            "description": "Session ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only return events with seq greater than this",
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of events to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionEvent"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Events retrieved"
          }
        },
        "tags": [
          "sessions"
        ]
      }
    }
  }
}